requirements.txt
aniso8601==9.0.1
attrs==22.1.0
click==8.1.3
coverage==6.4.4
fastjsonschema==2.16.2
Flask==2.2.2
Flask-RESTful==0.3.9
importlib-metadata==4.12.0
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
kafka-python==2.0.2
MarkupSafe==2.1.1
networkx==2.8.6
packaging==21.3
pluggy==1.0.0
py==1.11.0
pyparsing==3.0.9
pytest==7.1.3
pytz==2022.2.1
PyYAML==6.0
six==1.16.0
tomli==2.0.1
Werkzeug==2.2.2
zipp==3.8.1
